Add all entries section in primary tagging source details
