+++ src/newViews/TaggingFlow/PrimaryTagging/SourceDetails/styles.scss
@@ -1,0 +1,79 @@
+.source-preview {
+    display: flex;
+    flex-direction: column;
+
+    .tab-list {
+        flex-shrink: 0;
+
+        >* {
+            flex-grow: 1;
+        }
+    }
+
+    .simplified-tab {
+        display: flex;
+        align-items: center;
+        flex-grow: 1;
+        justify-content: center;
+        background-color: var(--dui-color-background-information);
+    }
+
+    .entry-list-tab {
+        flex-grow: 1;
+        background-color: var(--dui-color-background-information);
+        padding: calc(var(--dui-spacing-large) - var(--dui-spacing-medium));
+        overflow: auto;
+    }
+}
+
+// NOTE: styling for original tab is moved outside to make it work
+// for fullscreen mode
+.original-tab {
+    display: flex;
+    flex-direction: column;
+    flex-grow: 1;
+    background-color: var(--dui-color-background-information);
+    padding: calc(var(--dui-spacing-large) - var(--dui-spacing-small));
+
+    .actions {
+        display: flex;
+        flex-shrink: 0;
+        justify-content: flex-end;
+        padding: calc(var(--dui-spacing-medium) - var(--dui-spacing-small));
+
+        >* {
+            margin: var(--dui-spacing-small);
+        }
+    }
+
+    .content {
+        display: flex;
+        position: relative;
+        align-items: center;
+        flex-grow: 1;
+        justify-content: center;
+        margin: var(--dui-spacing-medium);
+        background-color: var(--dui-color-foreground);
+    }
+}
+
+.entry {
+    margin: var(--dui-spacing-medium);
+    border: var(--dui-width-separator-thin) solid var(--dui-color-accent);
+    background-color: var(--dui-color-foreground);
+    cursor: pointer;
+    padding: calc(var(--dui-spacing-medium) - var(--dui-spacing-small));
+
+    .excerpt {
+        padding: var(--dui-spacing-small);
+    }
+
+    .image {
+        max-width: 100%;
+    }
+
+    &.active {
+        border-right: var(--dui-width-separator-thick) solid var(--dui-color-accent);
+        background-color: var(--dui-color-background-information);
+    }
+}
